integrate longevity app into universe framework
