+++ tests/src/main/resources/universe-tests.properties
@@ -1,5 +1,5 @@
 corfu.cluster.name=static_cluster
 corfu.server.jar=build
 corfu.server.initialPort=9000
-server.version=0.3.0-SNAPSHOT
+server.version=0.3.1-SNAPSHOT
 test.data.clean=true
