+++ tests/src/test/java/org/corfudb/test/spec/DisconnectFirstServerSpec.java
@@ -99,7 +99,7 @@ public class DisconnectFirstServerSpec {
         log.info("**** First Insertion Verified... ****");
 
         // Get all nodes of cluster in separate variables
-        CorfuApplicationServer server0 = corfuCluster.getServerByIndex(0);
+        CorfuApplicationServer server0 = corfuCluster.getFirstServer();
         CorfuApplicationServer server1 = corfuCluster.getServerByIndex(1);
         CorfuApplicationServer server2 = corfuCluster.getServerByIndex(2);
