+++ tests/src/test/java/org/corfudb/test/spec/NodePausedAndPartitionedSpec.java
@@ -101,7 +101,7 @@ public class NodePausedAndPartitionedSpec {
         log.info("**** First Insertion Verified... ****");
 
         // Get all nodes of cluster in separate variables
-        CorfuApplicationServer server0 = corfuCluster.getServerByIndex(0);
+        CorfuApplicationServer server0 = corfuCluster.getFirstServer();
         CorfuApplicationServer server1 = corfuCluster.getServerByIndex(1);
         CorfuApplicationServer server2 = corfuCluster.getServerByIndex(2);
