+++ tests/src/test/java/org/corfudb/test/spec/TwoNodesDownSpec.java
@@ -109,7 +109,7 @@ public class TwoNodesDownSpec {
         log.info("First Insertion Verified...");
 
         //Should stop two nodes and then restart
-        CorfuApplicationServer server0 = corfuCluster.getServerByIndex(0);
+        CorfuApplicationServer server0 = corfuCluster.getFirstServer();
         CorfuApplicationServer server1 = corfuCluster.getServerByIndex(1);
         CorfuApplicationServer server2 = corfuCluster.getServerByIndex(2);
