+++ tests/src/test/java/org/corfudb/test/vm/stateful/ufo/poweroff/PowerOffOnThousandTimesInSingleNodeClusterTest.java
@@ -122,7 +122,7 @@ public class PowerOffOnThousandTimesInSingleNodeClusterTest extends AbstractCorf
             log.info("**** In Loop :: " + loopCount + " ****");
 
             log.info("**** PowerOff Single node ****");
-            VmCorfuServer vmServer = (VmCorfuServer) corfuCluster.getServerByIndex(0);
+            VmCorfuServer vmServer = (VmCorfuServer) corfuCluster.getFirstServer();
             vmServer.getVmManager().powerOff();
 
             log.info("**** PowerOn Single node ****");
