+++ tests/src/test/java/org/corfudb/test/vm/stateful/ufo/restart/RestartServiceThousandTimesInTwoNodeClusterTest.java
@@ -106,7 +106,7 @@ public class RestartServiceThousandTimesInTwoNodeClusterTest extends AbstractCor
         UfoUtils.verifyTableData(corfuStore, 0, count, manager, tableName, false);
         log.info("**** First Insertion Verified... ****");
 
-        CorfuApplicationServer server0 = corfuCluster.getServerByIndex(0);
+        CorfuApplicationServer server0 = corfuCluster.getFirstServer();
         log.info("**** Detach node server0 from cluster ****");
         ScenarioUtils.detachNodeAndVerify(corfuClient, server0, clientFixture);
