+++ universe/src/main/java/org/corfudb/universe/api/UniverseManager.java
@@ -2,15 +2,16 @@ package org.corfudb.universe.api;
 
 import lombok.Builder;
 import lombok.NonNull;
+import lombok.extern.slf4j.Slf4j;
 import org.corfudb.universe.api.universe.UniverseParams;
 import org.corfudb.universe.api.workflow.UniverseWorkflow.WorkflowConfig;
 import org.corfudb.universe.api.workflow.UniverseWorkflow.WorkflowContext;
 import org.corfudb.universe.infrastructure.docker.workflow.DockerUniverseWorkflow;
 import org.corfudb.universe.infrastructure.process.workflow.ProcessUniverseWorkflow;
 import org.corfudb.universe.infrastructure.vm.workflow.VmUniverseWorkflow;
-import org.corfudb.universe.scenario.fixture.Fixtures.UniverseFixture;
-import org.corfudb.universe.scenario.fixture.Fixtures.VmFixtureContext;
-import org.corfudb.universe.scenario.fixture.Fixtures.VmUniverseFixture;
+import org.corfudb.universe.scenario.fixture.UniverseFixture;
+import org.corfudb.universe.scenario.fixture.VmUniverseFixture;
+import org.corfudb.universe.scenario.fixture.VmUniverseFixture.VmFixtureContext;
 
 import java.util.function.Consumer;
 
@@ -18,6 +19,7 @@ import java.util.function.Consumer;
  * Manages UniverseWorkflow and provides api to build a universe workflow.
  */
 @Builder
+@Slf4j
 public class UniverseManager {
 
     @NonNull
@@ -36,6 +38,7 @@ public class UniverseManager {
                 .config(config)
                 .fixture(new UniverseFixture())
                 .build();
+
         DockerUniverseWorkflow wf = DockerUniverseWorkflow.builder()
                 .context(context)
                 .build();
@@ -43,8 +46,11 @@ public class UniverseManager {
         wf.init();
         try {
             action.accept(wf);
-        } finally {
-            wf.getContext().getUniverse().shutdown();
+        } catch (Exception ex) {
+            log.error("Universe error", ex);
+        }
+        finally {
+            wf.getUniverse().shutdown();
         }
 
         return wf;
@@ -70,7 +76,9 @@ public class UniverseManager {
         try {
             action.accept(wf);
         } finally {
-            wf.getContext().getUniverse().shutdown();
+            wf.getContext().getUniverse()
+                    .orElseThrow(() -> new IllegalStateException("Universe is not ready"))
+                    .shutdown();
         }
 
         return wf;
@@ -97,7 +105,9 @@ public class UniverseManager {
         try {
             action.accept(wf);
         } finally {
-            wf.getContext().getUniverse().shutdown();
+            wf.getContext().getUniverse()
+                    .orElseThrow(() -> new IllegalStateException("Universe is not ready"))
+                    .shutdown();
         }
 
         return wf;
