+++ universe/src/main/java/org/corfudb/universe/api/universe/group/cluster/Cluster.java
@@ -19,6 +19,6 @@ public interface Cluster<
     void bootstrap();
 
     enum ClusterType {
-        CORFU, PROM, CASSANDRA, MANGLE
+        CORFU, CORFU_LONGEVITY_APP, PROM, CASSANDRA, MANGLE
     }
 }
