+++ universe/src/main/java/org/corfudb/universe/infrastructure/docker/universe/group/cluster/DockerCorfuLongevityCluster.java
@@ -1,0 +1,69 @@
+package org.corfudb.universe.infrastructure.docker.universe.group.cluster;
+
+import com.spotify.docker.client.DockerClient;
+import lombok.Builder;
+import lombok.NonNull;
+import lombok.extern.slf4j.Slf4j;
+import org.corfudb.universe.api.common.LoggingParams;
+import org.corfudb.universe.api.deployment.docker.DockerContainerParams;
+import org.corfudb.universe.api.universe.UniverseParams;
+import org.corfudb.universe.api.universe.group.GroupParams.GenericGroupParams;
+import org.corfudb.universe.api.universe.group.cluster.AbstractCluster;
+import org.corfudb.universe.infrastructure.docker.DockerManager;
+import org.corfudb.universe.infrastructure.docker.universe.node.server.DockerServers.DockerCorfuLongevityApp;
+import org.corfudb.universe.universe.group.cluster.corfu.CorfuCluster;
+import org.corfudb.universe.universe.node.server.corfu.LongevityAppParams;
+
+/**
+ * Provides Docker implementation of {@link CorfuCluster}.
+ */
+@Slf4j
+public class DockerCorfuLongevityCluster extends AbstractCluster<
+        LongevityAppParams,
+        DockerContainerParams<LongevityAppParams>,
+        DockerCorfuLongevityApp,
+        GenericGroupParams<LongevityAppParams, DockerContainerParams<LongevityAppParams>>> {
+
+    @NonNull
+    private final DockerClient docker;
+
+    /**
+     * Support cluster
+     *
+     * @param docker          docker client
+     * @param containerParams cassandra server params
+     * @param universeParams  universe params
+     */
+    @Builder
+    public DockerCorfuLongevityCluster(
+            DockerClient docker, UniverseParams universeParams,
+            GenericGroupParams<LongevityAppParams, DockerContainerParams<LongevityAppParams>> containerParams,
+            LoggingParams loggingParams) {
+        super(containerParams, universeParams, loggingParams);
+        this.docker = docker;
+        init();
+    }
+
+    @Override
+    public void bootstrap() {
+        // NOOP
+    }
+
+    @Override
+    protected DockerCorfuLongevityApp buildServer(DockerContainerParams<LongevityAppParams> deploymentParams) {
+
+        DockerManager<LongevityAppParams> dockerManager = DockerManager
+                .<LongevityAppParams>builder()
+                .docker(docker)
+                .containerParams(deploymentParams)
+                .build();
+
+        return DockerCorfuLongevityApp.builder()
+                .containerParams(deploymentParams)
+                .groupParams(params)
+                .docker(docker)
+                .dockerManager(dockerManager)
+                .loggingParams(loggingParams)
+                .build();
+    }
+}
