+++ universe/src/main/java/org/corfudb/universe/infrastructure/vm/workflow/VmUniverseWorkflow.java
@@ -6,8 +6,10 @@ import org.corfudb.universe.api.common.LoggingParams;
 import org.corfudb.universe.api.workflow.UniverseWorkflow;
 import org.corfudb.universe.infrastructure.vm.universe.ApplianceManager;
 import org.corfudb.universe.infrastructure.vm.universe.VmUniverse;
-import org.corfudb.universe.scenario.fixture.Fixtures.VmFixtureContext;
-import org.corfudb.universe.scenario.fixture.Fixtures.VmUniverseFixture;
+import org.corfudb.universe.scenario.fixture.VmUniverseFixture.VmFixtureContext;
+import org.corfudb.universe.scenario.fixture.VmUniverseFixture;
+
+import java.util.Optional;
 
 @Builder
 public class VmUniverseWorkflow implements UniverseWorkflow<VmFixtureContext, VmUniverseFixture> {
@@ -37,7 +39,7 @@ public class VmUniverseWorkflow implements UniverseWorkflow<VmFixtureContext, Vm
                 .loggingParams(loggingParams)
                 .applianceManager(manager)
                 .build();
-        context.setUniverse(universe);
+        context.setUniverse(Optional.of(universe));
 
         context.setInitialized(true);
